made music player into service module
