+++ client/app/.env.dist
@@ -1,7 +1,11 @@
-AMPLIFY_AUTH_DISABLE=false
+AMPLIFY_AUTH_DISABLE=true
 AMPLIFY_AUTH_REGION=
 AMPLIFY_AUTH_USER_POOL_ID=
 AMPLIFY_AUTH_WEB_CLIENT_ID=
 CONCERT_START_TIME=
 WEBSOCKET_API_STUB=false
 WEBSOCKET_API_URL=
+MTURK_SONG_ID=
+MTURK_CHOICE_TYPE=
+MTURK_CHOICE_INTERVAL=20
+MTURK_CHOICE_TIMEOUT=5
